decimal value for block cycles
git-svn-id: https://svn.fao.org/projects/ruralinvest/RIV4/trunk@67711 d7fa552c-c32d-11e1-9f2d-6d39f2a56844

+++ RuralInvest4/src/main/java/riv/objects/FinanceMatrix.java
@@ -338,7 +338,7 @@ public class FinanceMatrix {
 		}
 	
 		for (BlockWithout block : project.getBlocksWithout()) {
-			int cycles = block.getCyclePerYear();
+			double cycles = block.getCyclePerYear();
 			// BLOCK: INCOME
 			for (BlockIncome income : block.getIncomes()) {
 				for (int i=0;i<project.getDuration();i++) {
@@ -399,7 +399,7 @@ public class FinanceMatrix {
 		}
 		
 		for (Block block : project.getBlocks()) {
-			int cycles = block.getCyclePerYear();
+			double cycles = block.getCyclePerYear();
 			// BLOCK: INCOME
 			for (BlockIncome income : block.getIncomes()) {
 				for (int i=0;i<project.getDuration();i++) {
